refactor index file
